+++ metautil.d.ts
@@ -49,16 +49,19 @@ export function sizeToBytes(size: string): number;
 export function namespaceByPath(namespace: object, path: string): object | null;
 
 type Every = {
-  month: number;
-  day: number;
+  MM: number;
+  DD: number;
+  wd: number;
   dd: number;
   hh: number;
   mm: number;
-  interval: number;
+  ms: number;
 };
 
 export type { Every };
 
+export function parseDay(s: string): number;
+export function parseMonth(s: string): number;
 export function parseEvery(s: string): Every;
 export function nextEvent(every: Every, date?: Date): number;
 export function makePrivate(instance: object): object;