+++ metautil.d.ts
@@ -48,6 +48,7 @@ export function sizeToBytes(size: string): number;
 export function namespaceByPath(namespace: object, path: string): object | null;
 
 type Every = {
+  YY: number;
   MM: number;
   DD: number;
   wd: number;