+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased][unreleased]
 
+- New functions: `parseDay` and `parseMonth`
+- Fixed and `parseEvery` and `nextEvent`, more tests cases
+
 ## [3.5.21][] - 2022-06-27
 
 - Update dependencies, eslint, and optimize package